fix: ci, module

+++ .github/workflows/linux-post.sh
@@ -10,13 +10,27 @@ cd ${BASE}
 export FUTURERESTORE_VERSION=$(git rev-list --count HEAD | tr -d '\n')
 export FUTURERESTORE_VERSION_RELEASE=$(cat version.txt | tr -d '\n')
 cd ${WORKFLOW_ROOT}
-echo "futurerestore-Linux-x86_64-${FUTURERESTORE_VERSION_RELEASE}-Build_${FUTURERESTORE_VERSION}-RELEASE.tar.xz" > name1.txt
-echo "futurerestore-Linux-x86_64-${FUTURERESTORE_VERSION_RELEASE}-Build_${FUTURERESTORE_VERSION}-DEBUG.tar.xz" > name2.txt
-echo "futurerestore-Linux-x86_64-${FUTURERESTORE_VERSION_RELEASE}-Build_${FUTURERESTORE_VERSION}-ASAN.tar.xz" > name3.txt
-cp -RpP "${TMPDIR}/Builder/linux_fix.sh" linux_fix.sh
-cp -RpP "${BASE}/cmake-build-release-x86_64/src/futurerestore" futurerestore
-tar cpPJvf "futurerestore1.tar.xz" futurerestore linux_fix.sh
+echo "futurerestore-Linux-x86_64-${FUTURERESTORE_VERSION_RELEASE}-Build_${FUTURERESTORE_VERSION}-RELEASE.tar.xz" >name1.txt
+echo "futurerestore-Linux-x86_64-${FUTURERESTORE_VERSION_RELEASE}-Build_${FUTURERESTORE_VERSION}-DEBUG.tar.xz" >name2.txt
+echo "futurerestore-Linux-x86_64-${FUTURERESTORE_VERSION_RELEASE}-Build_${FUTURERESTORE_VERSION}-ASAN.tar.xz" >name3.txt
+if [ -f "${TMPDIR}/Builder/linux_fix.sh" ]; then
+    cp -RpP "${TMPDIR}/Builder/linux_fix.sh" linux_fix.sh
+    cp -RpP "${BASE}/cmake-build-release-x86_64/src/futurerestore" futurerestore
+    tar cpPJvf "futurerestore1.tar.xz" futurerestore linux_fix.sh
+else
+    echo "Warning: linux_fix.sh not found, creating archive without it"
+    cp -RpP "${BASE}/cmake-build-release-x86_64/src/futurerestore" futurerestore
+    tar cpPJvf "futurerestore1.tar.xz" futurerestore
+fi
 cp -RpP "${BASE}/cmake-build-debug-x86_64/src/futurerestore" futurerestore
-tar cpPJvf "futurerestore2.tar.xz" futurerestore linux_fix.sh
+if [ -f "linux_fix.sh" ]; then
+    tar cpPJvf "futurerestore2.tar.xz" futurerestore linux_fix.sh
+else
+    tar cpPJvf "futurerestore2.tar.xz" futurerestore
+fi
 cp -RpP "${BASE}/cmake-build-asan-x86_64/src/futurerestore" futurerestore
-tar cpPJvf "futurerestore3.tar.xz" futurerestore linux_fix.sh
+if [ -f "linux_fix.sh" ]; then
+    tar cpPJvf "futurerestore3.tar.xz" futurerestore linux_fix.sh
+else
+    tar cpPJvf "futurerestore3.tar.xz" futurerestore
+fi
